Move grid test to proper source set
