+++ app/build.gradle.kts
@@ -35,4 +35,6 @@ dependencies {
     // viewModel helpers
     implementation("androidx.lifecycle:lifecycle-runtime-compose:2.8.1")
     implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.8.1")
+
+    testImplementation("junit:junit:4.13.2")
 }
